+++ CHANGELOG.md
@@ -9,6 +9,7 @@
 * [#587](https://github.com/plotly/dash-bio/pull/587) Added JSME component.
 * [#628](https://github.com/plotly/dash-bio/pull/628) Added option to add colored labels to rows and columns on Clustergram.
 * [#640](https://github.com/plotly/dash-bio/pull/640) Added loading states to each component to support custom loading css with `dash-data-is-loading`.
+* [#642](https://github.com/plotly/dash-bio/pull/642) Added `scrollZoom` property to Molecule2dViewer.
 
 
 ### Changed